Reduce duplication between controller impls
